added default value at pagination

+++ src/building-blocks/BuildingBlocks.Persistence.EFCore/Parameters/GetListParamaters.cs
@@ -6,6 +6,6 @@ public sealed class GetListParameters<TEntity> : ReadParameters {
 	public Expression<Func<TEntity, Boolean>>? Predicate { get; set; }
 	public Func<IQueryable<TEntity>, IOrderedQueryable<TEntity>>? OrderBy { get; set; }
 	public Func<IQueryable<TEntity>, IIncludableQueryable<TEntity, Object>>? Include { get; set; }
-	public Int32 Index { get; set; }
-	public Int32 Size { get; set; }
+	public UInt32 Index { get; set; }
+	public UInt64 Size { get; set; }
 }
